+++ crates/captval_derive/tests/expand/complete_struct.rs
@@ -1,0 +1,13 @@
+use captval_derive::Captval;
+
+#[derive(Debug, Captval)]
+pub struct ContactForm {
+    name: String,
+    #[allow(dead_code)]
+    phone: String,
+    email: String,
+    #[allow(dead_code)]
+    message: String,
+    #[captcha]
+    token: String,
+}